Enclose all retorts within a div
This will help users that want to place the retorts on a separate row above the post actions. This is especially helpful if your post has at least a handful of retorts. It’s also very handy when using a mobile device, where a single retort makes the full action bar not fit within the screen.